Draft: Pr/fix ci (#39)
* ci: only use clang-format 14

9 and 14 have inconsistent formatting, and it's not clear from dumping
the llvm style what changed or how to modify the config file to force
consistency. Simpler to standardize on one version and 14 is included
with Ubuntu 22.04 lts so seems like a good version to use.

* ci: don't run tests on CPU

Ideally we would have this working, but more important to have a CI
test that is passing and we can enforce, and building on oneAPI is
such a test and is a reasonable starting point.