pow_complex test: use matcher, workaround oneAPI bug
The std::pow call on certain inputs produces (nan, nan) instead
of (0, 0), which is produced by all other implementations and
by std::exp(y * log(x)). This PR makes use of the new matcher,
which was used to track down actual problematic cases, and adds
a workaround that is enabled by macro definition only when
IntelLLVM compiler type is detected.